More variable life

+++ src/simulation/Simulation.h
@@ -82,7 +82,7 @@ public:
 	int GSPEED;
 	unsigned char gol[YRES][XRES];
 	unsigned short gol2[YRES][XRES][9];
-	unsigned short golvari[YRES][XRES];
+	unsigned int golvari[YRES][XRES];
 	//Air sim
 	float (*vx)[XRES/CELL];
 	float (*vy)[XRES/CELL];
